requirement.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
opencv-contrib-python-headless>=4.9.0
numpy>=1.26.0
scikit-image>=0.22.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
sqlalchemy>=2.0.25